Add files via upload
Create a structure named “Point” and 2 data members: X and Y coordinate. Declare 2 points: P1 and P2. Determine if P2 is to the right or left of P1 or on same axis , by comparing the x coordinates. ( if p1.x is more than p2.x , it is to the right )